refactor(handlers): split proxy role helpers
